feat(scooters): added publish scooter feature

+++ lib/scooters/domain/dto/scooter_request.dto.dart
@@ -5,7 +5,7 @@ class ScooterRequestDTO {
   String? neighborhood;
   String? city;
   String? district;
-  int? price;
+  double? price;
   String? model;
   int? profileId;
